Restructure ensemble evaluator
