+++ src/ert/ensemble_evaluator/__init__.py
@@ -1,13 +1,22 @@
-from ._builder import (
-    Ensemble,
+from ._ensemble import (
     ForwardModelStep,
     Realization,
 )
+from ._ensemble import (
+    LegacyEnsemble as Ensemble,
+)
+from ._wait_for_evaluator import wait_for_evaluator
 from .config import EvaluatorServerConfig
 from .evaluator import EnsembleEvaluator
 from .event import EndEvent, FullSnapshotEvent, SnapshotUpdateEvent
 from .monitor import Monitor
-from .snapshot import PartialSnapshot, Snapshot
+from .snapshot import (
+    ForwardModel,
+    PartialSnapshot,
+    RealizationSnapshot,
+    Snapshot,
+    SnapshotDict,
+)
 
 __all__ = (
     "EndEvent",
@@ -18,7 +27,11 @@ __all__ = (
     "FullSnapshotEvent",
     "Monitor",
     "PartialSnapshot",
-    "Realization",
     "Snapshot",
     "SnapshotUpdateEvent",
+    "wait_for_evaluator",
+    "RealizationSnapshot",
+    "SnapshotDict",
+    "ForwardModel",
+    "Realization",
 )
