--- src/ert/ensemble_evaluator/_builder/__init__.py
@@ -1,8 +1,0 @@
-from ._legacy import LegacyEnsemble as Ensemble
-from ._realization import ForwardModelStep, Realization
-
-__all__ = (
-    "Ensemble",
-    "ForwardModelStep",
-    "Realization",
-)
